refactor(session-notification): extract platform functions to comply with 200 LOC rule
